requirements.txt
mysqlclient==1.3.4
Flask
gunicorn
pytest==4.3.0
pytest-cov
flake8
flask-debugtoolbar
Flask-WTF
Flask-SQLAlchemy
Flask-Login
WTForms-Components
Click
sqlalchemy-utils
bcrypt
sendgrid
redis
celery
python-dotenv
flask-moment
Faker
flask-oauthlib==0.9.5
